+++ data/src/main/java/com/roacult/kero/oxxy/projet2eme/network/entities/ProfileEntities.kt
@@ -1,3 +1,3 @@
 package com.roacult.kero.oxxy.projet2eme.network.entities
 
-data class UserId(val id:Long)
+data class UserId(val id:Int)
